requirements.txt
Pillow==10.0.1
yt-dlp==2023.7.6